meson: Fix removal of static lib
[why]
With commit b66ee67e1 the sttaic lib has been removed from the Meson
build description (the linking of the lib).
But the static link dependency is still there and can not be filled.
This means Meson builds are impossible since that commit.

[how]
Also remove the dependency variable for the static library.

[note]
Both is only commented out, because for me the reason is not so clear.
What would be the problem with a static lib?

And then we have this comment in the README file:

    We need to have static library, because the code for auth and
    client_* depends on getuid, getid functions, that will be
    implemented later.
    On windows this possible to do (at least to my knowledge) only
    with static library

This means we will need the static library later again?

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>