Remove storage class register from interface
[why]
With C++17 the storage class register has been dropped. It has always
been only a 'hint' to the compiler, a 'wish'. While the register storage
class is still valid in C, it is impossible to include header files that
use that storage class in c++ sources that use c++17 or later.

Doocs is now a C++17 source, so we need to adapt this.

[how]
Just drop the storage class modifier. Is has been ignored before anyway.

[note]
Also fix wrong decumentation of  clnttcp_creat() in clnt.h, where the
docs said 'register' but the original function definition did not.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>